Refactor: Integrate Gemini API and Browser Tool
This commit refactors the ai-slicer project to use the Google Gemini API
instead of OpenAI. It also introduces a 'Browser as a Tool' functionality,
allowing the assistant to fetch content from local network URLs.

Key changes:
- Updated requirements.txt with google-generativeai, requests, and markdownify.
- Added `fetch_local_url_content` tool in `tools.py` to get content from URLs.
- Refactored `assist.py`:
    - Replaced OpenAI client with Gemini client (`gemini-1.5-flash-latest`).
    - Defined Gemini tools programmatically (for `fetch_local_url_content`).
    - Updated the system prompt to "Arturo" and incorporated the new tool's usage.
    - Modified `ask_question_memory` to use Gemini's API for chat and function calling.
    - Commented out OpenAI-based TTS functions.
- Updated `ai-slicer.py` to comment out calls to the now-removed TTS functions.

Manual testing is required to ensure the Gemini API key is configured correctly
and the new browser tool functions as expected with local network resources.
